+++ crates/ruff_python_parser/src/lib.rs
@@ -80,7 +80,6 @@ use ruff_text_size::{Ranged, TextRange};
 
 mod error;
 pub mod lexer;
-pub mod parenthesize;
 mod parser;
 pub mod semantic_errors;
 mod string;
